isolate types to single types file
