--- src/interfaces/client-options.ts
@@ -1,10 +1,0 @@
-import { Cache } from './cache';
-import { Exchange } from './exchange';
-
-export interface ClientOptions {
-  url: string;
-  fetchOptions?: object | (() => object);
-  cache?: Cache;
-  initialCache?: object;
-  transformExchange?: (IExchange, IClient) => Exchange;
-}
