--- src/interfaces/error.ts
@@ -1,3 +1,0 @@
-export interface GraphQLError {
-  message?: string;
-}
