--- src/interfaces/index.ts
@@ -1,10 +1,0 @@
-export { IClient } from './client';
-export { Mutation } from './mutation';
-export { Query } from './query';
-export { ClientOptions } from './client-options';
-export { Cache } from './cache';
-export { Exchange, ExchangeResult } from './exchange';
-export { GraphQLError } from './error';
-export { ClientEventType, EventFn } from './events';
-export { Operation } from './operation';
-export { SubscriptionObserver, Subscription } from './subscription';
