--- src/interfaces/mutation.ts
@@ -1,5 +1,0 @@
-import { Query } from './query';
-
-export interface Mutation {
-  [key: string]: Query;
-}
