--- src/interfaces/query.ts
@@ -1,4 +1,0 @@
-export interface Query {
-  query: string;
-  variables?: object;
-}
