--- src/interfaces/subscription.ts
@@ -1,11 +1,0 @@
-// These are *not* Observable primitives, but are instead intended to
-// be used with the subscriptionExchange
-
-export interface SubscriptionObserver {
-  next: (IExecutionResult) => void;
-  error: (Error) => void;
-}
-
-export interface Subscription {
-  unsubscribe: () => void;
-}
